+++ modules/core.lua
@@ -2,5 +2,5 @@
 EZOArmory = EZOArmory or {}
 
 EZOArmory.ADDON_NAME    = "EZOArmory"
-EZOArmory.ADDON_VERSION = "0.1.10"
+EZOArmory.ADDON_VERSION = "0.2.0"
 EZOArmory.AUTHOR        = "@Zuriplayer"